Declare Wear OrderItem type with dedicated parsing
