+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/model/Refund.kt
@@ -148,11 +148,3 @@ fun List<Refund>.getMaxRefundQuantities(
     }
     return map
 }
-
-data class OrderItem(
-    val itemId: Long,
-    val quantity: Float,
-    val totalTax: BigDecimal,
-    val price: BigDecimal,
-    val total: BigDecimal
-)
